del C / del w and weights and biases update
Added a way to calculate del C / del w then added a way for the weights and biases to be updated. Tested this out by making it try to learn a XOR gate and it works for all inputs but 1, 1 which sometimes gets it right and sometimes doesn't, so I will need to look into that.