Foward propagation
Added a vectorized implemintation of foward propagation that does all the training examples at the same time, changed the activation functions as I am still trying to learn about all of them and deciding which to use. Made the network architecture much bigger than it needs to be to test if this implemintation is scalable.